requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
future==0.18.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pbr==5.4.4
pipenv==2018.11.26
pylint==2.4.4
python-twitter==3.5
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
stevedore==1.31.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wrapt==1.11.2
